compiling denort
