+++ cli/args/mod.rs
@@ -58,8 +58,8 @@ use deno_core::serde_json;
 use deno_core::url::Url;
 use deno_graph::GraphKind;
 pub use deno_json::check_warn_tsconfig;
+use deno_lib::args::has_flag_env_var;
 use deno_lib::cache::DenoDirProvider;
-use deno_lib::env::has_flag_env_var;
 use deno_lib::worker::StorageKeyResolver;
 use deno_lint::linter::LintConfig as DenoLintConfig;
 use deno_npm::npm_rc::NpmRc;
@@ -606,147 +606,6 @@ pub fn create_default_npmrc() -> Arc<ResolvedNpmRc> {
     registry_configs: Default::default(),
   })
 }
-
-#[derive(Error, Debug, Clone, deno_error::JsError)]
-#[class(generic)]
-pub enum RootCertStoreLoadError {
-  #[error(
-    "Unknown certificate store \"{0}\" specified (allowed: \"system,mozilla\")"
-  )]
-  UnknownStore(String),
-  #[error("Unable to add pem file to certificate store: {0}")]
-  FailedAddPemFile(String),
-  #[error("Failed opening CA file: {0}")]
-  CaFileOpenError(String),
-}
-
-/// Create and populate a root cert store based on the passed options and
-/// environment.
-pub fn get_root_cert_store(
-  maybe_root_path: Option<PathBuf>,
-  maybe_ca_stores: Option<Vec<String>>,
-  maybe_ca_data: Option<CaData>,
-) -> Result<RootCertStore, RootCertStoreLoadError> {
-  let mut root_cert_store = RootCertStore::empty();
-  let ca_stores: Vec<String> = maybe_ca_stores
-    .or_else(|| {
-      let env_ca_store = env::var("DENO_TLS_CA_STORE").ok()?;
-      Some(
-        env_ca_store
-          .split(',')
-          .map(|s| s.trim().to_string())
-          .filter(|s| !s.is_empty())
-          .collect(),
-      )
-    })
-    .unwrap_or_else(|| vec!["mozilla".to_string()]);
-
-  for store in ca_stores.iter() {
-    match store.as_str() {
-      "mozilla" => {
-        root_cert_store.extend(webpki_roots::TLS_SERVER_ROOTS.to_vec());
-      }
-      "system" => {
-        let roots = load_native_certs().expect("could not load platform certs");
-        for root in roots {
-          if let Err(err) = root_cert_store
-            .add(rustls::pki_types::CertificateDer::from(root.0.clone()))
-          {
-            log::error!(
-              "{}",
-              colors::yellow(&format!(
-                "Unable to add system certificate to certificate store: {:?}",
-                err
-              ))
-            );
-            let hex_encoded_root = faster_hex::hex_string(&root.0);
-            log::error!("{}", colors::gray(&hex_encoded_root));
-          }
-        }
-      }
-      _ => {
-        return Err(RootCertStoreLoadError::UnknownStore(store.clone()));
-      }
-    }
-  }
-
-  let ca_data =
-    maybe_ca_data.or_else(|| env::var("DENO_CERT").ok().map(CaData::File));
-  if let Some(ca_data) = ca_data {
-    let result = match ca_data {
-      CaData::File(ca_file) => {
-        let ca_file = if let Some(root) = &maybe_root_path {
-          root.join(&ca_file)
-        } else {
-          PathBuf::from(ca_file)
-        };
-        let certfile = std::fs::File::open(ca_file).map_err(|err| {
-          RootCertStoreLoadError::CaFileOpenError(err.to_string())
-        })?;
-        let mut reader = BufReader::new(certfile);
-        rustls_pemfile::certs(&mut reader).collect::<Result<Vec<_>, _>>()
-      }
-      CaData::Bytes(data) => {
-        let mut reader = BufReader::new(Cursor::new(data));
-        rustls_pemfile::certs(&mut reader).collect::<Result<Vec<_>, _>>()
-      }
-    };
-
-    match result {
-      Ok(certs) => {
-        root_cert_store.add_parsable_certificates(certs);
-      }
-      Err(e) => {
-        return Err(RootCertStoreLoadError::FailedAddPemFile(e.to_string()));
-      }
-    }
-  }
-
-  Ok(root_cert_store)
-}
-
-/// State provided to the process via an environment variable.
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub struct NpmProcessState {
-  pub kind: NpmProcessStateKind,
-  pub local_node_modules_path: Option<String>,
-}
-
-#[derive(Clone, Debug, Serialize, Deserialize)]
-pub enum NpmProcessStateKind {
-  Snapshot(deno_npm::resolution::SerializedNpmResolutionSnapshot),
-  Byonm,
-}
-
-static NPM_PROCESS_STATE: Lazy<Option<NpmProcessState>> = Lazy::new(|| {
-  use deno_runtime::ops::process::NPM_RESOLUTION_STATE_FD_ENV_VAR_NAME;
-  let fd = std::env::var(NPM_RESOLUTION_STATE_FD_ENV_VAR_NAME).ok()?;
-  std::env::remove_var(NPM_RESOLUTION_STATE_FD_ENV_VAR_NAME);
-  let fd = fd.parse::<usize>().ok()?;
-  let mut file = {
-    use deno_runtime::deno_io::FromRawIoHandle;
-    unsafe { std::fs::File::from_raw_io_handle(fd as _) }
-  };
-  let mut buf = Vec::new();
-  // seek to beginning. after the file is written the position will be inherited by this subprocess,
-  // and also this file might have been read before
-  file.seek(std::io::SeekFrom::Start(0)).unwrap();
-  file
-    .read_to_end(&mut buf)
-    .inspect_err(|e| {
-      log::error!("failed to read npm process state from fd {fd}: {e}");
-    })
-    .ok()?;
-  let state: NpmProcessState = serde_json::from_slice(&buf)
-    .inspect_err(|e| {
-      log::error!(
-        "failed to deserialize npm process state: {e} {}",
-        String::from_utf8_lossy(&buf)
-      )
-    })
-    .ok()?;
-  Some(state)
-});
 
 /// Overrides for the options below that when set will
 /// use these values over the values derived from the
@@ -1990,15 +1849,6 @@ fn resolve_import_map_specifier(
 /// Resolves the no_prompt value based on the cli flags and environment.
 pub fn resolve_no_prompt(flags: &PermissionFlags) -> bool {
   flags.no_prompt || has_flag_env_var("DENO_NO_PROMPT")
-}
-
-pub fn npm_pkg_req_ref_to_binary_command(
-  req_ref: &NpmPackageReqReference,
-) -> String {
-  req_ref
-    .sub_path()
-    .map(|s| s.to_string())
-    .unwrap_or_else(|| req_ref.req().name.to_string())
 }
 
 pub fn config_to_deno_graph_workspace_member(
