+++ cli/emit.rs
@@ -20,9 +20,9 @@ use deno_error::JsErrorBox;
 use deno_graph::MediaType;
 use deno_graph::Module;
 use deno_graph::ModuleGraph;
+use deno_lib::util::hash::FastInsecureHasher;
 
 use crate::cache::EmitCache;
-use crate::cache::FastInsecureHasher;
 use crate::cache::ParsedSourceCache;
 use crate::resolver::CliCjsTracker;
