+++ cli/lib/lib.rs
@@ -1,7 +1,8 @@
 // Copyright 2018-2025 the Deno authors. MIT license.
 
+pub mod args;
 pub mod cache;
-pub mod env;
+pub mod loader;
 pub mod npm;
 pub mod shared;
 pub mod standalone;
