+++ cli/lib/util/hash.rs
