+++ cli/npm/installer/global.rs
@@ -9,6 +9,7 @@ use async_trait::async_trait;
 use deno_core::futures::stream::FuturesUnordered;
 use deno_core::futures::StreamExt;
 use deno_error::JsErrorBox;
+use deno_lib::util::hash::FastInsecureHasher;
 use deno_npm::NpmResolutionPackage;
 use deno_npm::NpmSystemInfo;
 use deno_resolver::npm::managed::NpmResolutionCell;
@@ -17,7 +18,6 @@ use super::common::lifecycle_scripts::LifecycleScriptsStrategy;
 use super::common::NpmPackageFsInstaller;
 use super::PackageCaching;
 use crate::args::LifecycleScriptsConfig;
-use crate::cache::FastInsecureHasher;
 use crate::colors;
 use crate::npm::CliNpmCache;
 use crate::npm::CliNpmTarballCache;
